Moved the `PostProcessHelper` and used it in a module to cover more cases where it wouldn't use the correct spanfactory.
