+++ console-framework-client/src/main/java/io/axoniq/console/framework/AxoniqConsoleConfigurerModule.java
@@ -246,6 +246,7 @@ public class AxoniqConsoleConfigurerModule implements ConfigurerModule {
         });
 
         new AxoniqConsoleAggregateConfigurerModule().configureModule(configurer);
+        new AxoniqConsoleEnhancingConfigurerModule(spanMatcherPredicateMap).configureModule(configurer);
     }
 
     /**
